Run tests headful, with xvfb in CI
After experimenting and debugging, it looks like multisampled depth
buffers in headless Chrome (with no GPU) is buggy and it fails with the
error:

WebGL: INVALID_OPERATION: renderbufferStorageMultisample: samples out of
range

It seems to work when running headful though. Note: ignoreDefaultArgs:
true on the puppeteer.launch options doesn't fix this either, so it
doesn't seem to be a flag that's being set (e.g. --disable-gpu).

Running with xvfb in the CI (the primary location for running the tests)
should fix this.